% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{awards}
\alias{awards}
\title{EU State Aid Awards}
\format{
A tibble with X variables and X observations:

\describe{

\item{\code{key_ID}}{Numeric. An ID number that uniquely identifies all
observations.}

}
}
\usage{
cases
}
\description{
A dataset of all awards granted by EU member states to firms
as part of approved state aid schemes from 2016 through 2020.
}
\details{
The State Aid Modernisation programme (SAM) was adopted in 2014
and came into force on 1 July 2016.
}
\keyword{datasets}
