% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awards_csts_bt.R
\docType{data}
\name{awards_csts_bt}
\alias{awards_csts_bt}
\title{Award-level cross-sectional time-series data by beneficiary type}
\format{
A data frame with 9 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{\code{numeric}. The year the case was opened by the
Commission.}
\item{member_state_id}{\code{numeric}. An ID number that uniquely
identifies each member state. This ID number is assigned when member states
are sorted by accession date and then alphabetically. }
\item{member_state}{\code{string}. The name of the member state that the
Commission opened the case against. }
\item{member_state_code}{\code{string}. A two letter code assigned by the
Commission that uniquely identifies each member state. }
\item{beneficiary_type_id}{\code{numeric}. An ID number that uniquely
identifies each type of beneficiary. Coded \code{1} for \code{Small or
medium-sized enterprise (SME)} and \code{2} for \code{Large enterprise}.}
\item{beneficiary_type}{\code{string}. The type of the beneficiary of the
state aid award. Coded \code{Small or medium-sized enterprise (SME)} or
\code{Large enterprise}. Uses the Commission’s standard definition of small
and median-sized enterprises (SMEs). }
\item{count_awards}{\code{numeric}. A count of the number of state aid
awards granted by the member state at this level of aggregation.}
\item{total_amount_awards}{\code{numeric}. The total amount of awards
granted at this level of aggregation in euros.}
}
}
\usage{
awards_csts_bt
}
\description{
This dataset includes aggregated data on the number of state aid awards per
member state per year (cross-sectional time-series data) broken down by
beneficiary type. There is one observation per member state per year per
beneficiary type (2016-2020).
}
\keyword{datasets}
