% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awards_csts_ns.R
\docType{data}
\name{awards_csts_ns}
\alias{awards_csts_ns}
\title{CSTS data on cases (by member state and NACE sector)}
\format{
A data frame with 10 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the case was opened by the Commission.}
\item{member_state_id}{Numeric. An ID number that uniquely identifies each
member state. This ID number is assigned when member states are sorted by
accession date and then alphabetically. }
\item{member_state}{String. The name of the member state that the
Commission opened the case against. }
\item{member_state_code}{String. A two letter code assigned by the
Commission that uniquely identifies each member state. }
\item{nace_sector_id}{Numeric. An ID number that uniquely identifies each
NACE sector.}
\item{nace_sector}{String. The NACE sector for the state aid award.}
\item{nace_sector_code}{String. The NACE sector code for the state aid
award (a single capital letter).}
\item{count_awards}{Numeric. A count of the number of state aid awards
granted by the member state at this level of aggregation.}
\item{total_amount_awards}{Numeric. The total amount of awards granted at
this level of aggregation in euros.}
}
}
\usage{
awards_csts_ns
}
\description{
This dataset includes aggregated data on the number of state aid awards per
member state per year (cross-sectional time-series data) broken down by
NACE sector. There is one observation per member state per year per NACE
sector (2016-2020).
}
\keyword{datasets}
