% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{cases}
\alias{cases}
\title{EU State Aid Cases}
\format{
A tibble with 33 variables and 33,715 observations:
\describe{

\item{\code{key_ID}}{Numeric. An ID number that uniquely identifies all
observations.}

\item{\code{case_number}}{String. An ID number that uniquely identifies each
case.}

\item{\code{procedure_number}}{String. A list of ID numbers that indentify
the proceedures associated with the case, separated by a comma.}

\item{\code{member_state}}{String. The name of the member state that notified
that state aid measure.}

\item{\code{directorate_general}}{String. The name of the Directorate-General
of the Commission that opened the case against the member state. Possible
values include: \itemize{\item{Agriculture and Rural Development},
\item{Competition}, and \item{Maritime Affairs and Fisheries}.}}

\item{\code{directorate_general_code}}{String. The alphabetical code for the
Directorate-General of the Commission that opened the case against the member
state. Possible values include: \itemize{\item{\code{AGRI} (for DG
Agriculture and Rural Development)}, \item{\code{COMP} (for DG Competition)},
and \item{\code{MARE} (for DG Maritime Affairs and Fisheries)}.}}

\item{\code{case_type}}{String. The type of the case. Possible values
include: \itemize{\item{\code{ad hoc case}} \item{\code{individual
application}} \item{\code{scheme}} \item{\code{not applicable}}}}

\item{\code{notification_date}}{String. The date the state aid measure was
notified to the Commission under Article 108(3) TFEU in the format
\code{YYYY-MM-DD.}}

\item{\code{decision_date}}{String. The date of the last decision made by the
Commission in the case in the format \code{YYYY-MM-DD}. Coded \code{NA} if
the measure is exempt from notification under Article 108(4) TFEU.}

\item{\code{decisions}}{String. A list of decisions made by the Commission in
the case in chronological order, separated by a semi-colon. Possible values
include: \itemize{
\item{\code{Article 4(2) of Council Regulation (EC) No
659/1999: decision does not constitute aid} — During its preliminary
investigation, Commission decided that the state aid measure notified by the
member state does not constitute state aid under Article 107 of the Treaty on
the Functioning of the European Union (TFEU).}
\item{\code{Article 4(3) of Council
Regulation (EC) No 659/1999: decision not to raise objections} — During its
preliminary investigation, Commission decided not to raise objections to the
state aid measure notified by the member state because the measure is
compatible with the rules of the single market, consistent with Article 107
of the Treaty on the Functioning of the European Union (TFEU).}
}}

\item{\code{count_decisions}}{Numeric. The number of decisions made by the
Commission in the case. Coded \code{NA} if the state aid measure was exempt
from notification under Article 108(3) of the Treaty on the Functioning of
the European Union.}

\item{\code{outcome}}{String. The outcome of the case. Possible values
include: \itemize{ \item{\code{exempt from notification} — The measure
notified by the member state is exempt from notification under Article 108(3)
TFEU so the Commission did not conduct a preliminary investigation.}
\item{\code{does not constitute aid} — Either during its preliminary
investigation or after a formal investigation, the Commission decided that
the measure notified by the member state does not constitute state aid under
Article 107 TFEU.} \item{\code{no objection} — During its preliminary
investigation, the Commisison decided that it had no objection to the state
aid measure because the measure is compatible with the rules of the single
market, as defined in Article 107 TFEU.} \item{\code{positive decision} —
After a formal investigation, the Commission decided that the state aid
measure is compatible with the rules of the single market, as defined in
Article 107 TFEU.} \item{\code{negative decision} — After a formal
investigation, the Commission decided that the state aid measure is not
compatible with the rules of the single market, as defined in Article 107
TFEU.} \item{\code{conditional decision} — After a formal investigation, the
Commission decided to approve the state aid measure subject to conditions
to make the measure compatible with the rules of the single market,
as defined in Article 107 TFEU.} \item{\code{notification withdrawn} — The
member state that notified the state aid measure withdrew its notification,
either during the Commission's preliminary investigation or during a formal
investigation.} \item{\code{missing record} — The
Commission initiated a formal investigation but the database of state aid
cases managed by DG Competition does not record the Commission making any
decisions after initiating the formal investigation.} }}

\item{\code{outcome_phase_1}}{String. The outcome of phase 1 of the state aid
procedure (the preliminary examination).}

\item{\code{outcome_phase_2}}{String. The outcome of phase 2 of the state aid
procedure (the formal investigation). Coded \code{not applicable} if the
Commission did not initiate a formal investigation under Article 7 of Council
Regulation (EC) No 659/1999.}

\item{\code{exempt}}{Numeric. A dummy variable indicating whether the state
aid measure is exempt from the Article 108(3) notification requirement under
Article 108(4) TFEU.}

\item{\code{preliminary_investigation}}{Numeric. A dummy variable indicating
whether the Commission initiated a preliminary examination under Article
108(2) TFEU.}

\item{\code{formal_investigation}}{Numeric. A dummy variable indicating
whether the Commission initiated a formal investigation under Article 7 of
Council Regulation (EC) No 659/1999.}

\item{\code{no_objection}}{Numeric. A dummy variable indicating whether the
Commission decided not to raise an objection under Article 4(3) of Council
Regulation (EC) No 659/1999.}

\item{\code{not_aid}}{Numeric. A dummy variable indicating whether the
Commission decided that the measure does not constitute aid under Article 107
TFEU, either after the preliminary examination under Article 4(2) of Council
Regulation (EC) No 659/1999, or after a formal investigation under Article
7(2) of Council Regulation (EC) No 659/1999.}

\item{\code{positive}}{Numeric. A dummy variable indicating whether the
Commission decided that the measure is compatible with the rules of the
single market, as defined in Article 107 TFEU.}

\item{\code{negative}}{Numeric. A dummy variable indicating whether the
Commission decided that the measure is not compatible with the rules of the
single market, as defined in Article 107 TFEU.}

\item{\code{conditional}}{Numeric. A dummy variable indicating whether the
Commission decided to approve the state aid measure subject to conditions to
make the measure compatible with the rules of the single market, as defined
in Article 107 TFEU.}

\item{\code{withdrawal}}{Numeric. A dummy variable indicating whether the
member state withdrew the notification, either before the formal investigaton
procedure, under Article 8(1) of Council Regulation (EC) No 659/1999, or
after the formal investigation procedure, under Article 8(2) of Council
Regulation (EC) No 659/1999.}

\item{\code{referral}}{Numeric. A dummy variable indicating whether the
Commission referred the case to the Court of Justice for non-compliance with
a decision under Article 23(1) of Councl Regulation (EC) No 659/1999, for
non-implementation of a Commission decision under Article 108(2) (ex Article
88(2)) of the Treaty on the Functioning of the European Union (TFEU), for
non-compliance with a Court judgment under Article 260(2) (ex Article 228(2))
of the Treaty on the Functioning of the European Union (TFEU), or for
non-compliance with an injunction under Article 12 of Council Regulation (EC)
No 659/1999.}

\item{\code{recovery}}{Numeric. A dummy variable indicating whether the
Commission decided that the measure is not compatible with the rules of the
single market, as defined in Article 107 TFEU, and decided that the member
state must take all necessary measures to recover the aid from the
beneficiary.}

\item{\code{start_date}}{String. The start date of the state aid measure in
the format \code{YYYY-MM-DD}. Coded \code{NA} if not applicable.}

\item{\code{start_date}}{String. The end date of the state aid measure
in the format \code{YYYY-MM-DD}. Coded \code{NA} if not applicable.}

\item{\code{aid_instruments}}{String. A list of instruments used by the state
aid measure, separated by a comma. Coded \code{NA} if no specific instruments
are reported. Possible values include:
\itemize{
\item{\code{debt write-off} — Definition.}
\item{\code{direct grant} — Definition.}
\item{\code{direct grant/interest rate subsidy} — Definition.}
\item{\code{fiscal measure} — Definition.}
\item{\code{guarantee} — Definition.}
\item{\code{interest subsidy} — Definition.}
\item{\code{loan/repayable advances} — Definition.}
\item{\code{other forms of equity intervention} — Definition.}
\item{\code{other forms of tax advantage} — Definition.}
\item{\code{provision of risk capital} — Definition.}
\item{\code{provision of risk finance} — Definition.}
\item{\code{reduction of social security contributions} — Definition.}
\item{\code{reimbursable grant} — Definition.}
\item{\code{repayable advances} — Definition.}
\item{\code{soft loan} — Definition.}
\item{\code{subsidised services} — Definition.}
\item{\code{tax advantage or tax exemption} — Definition.}
\item{\code{tax allowance} — Definition.}
\item{\code{tax base reduction} — Definition.}
\item{\code{tax deferment} — Definition.}
\item{\code{tax rate reduction} — Definition.}
}}

\item{\code{count_aid_instruments}}{Numeric. The number of instruments used
by the state aid measure.}

\item{\code{NACE_codes}}{String. A list of NACE codes associated with the
state aid measure, separated by a comma. Coded \code{NA} if no specific NACE
codes are reported.}

\item{\code{count_NACE_codes}}{Numeric. The number of NACE codes associated
with the state aid measure.}

\item{\code{OJ_citations}}{String. A list of citations in the Official
Journal (OJ) for decisions associated with the case, separated by a comma.}

\item{\code{count_OJ_citations}}{Numeric. The number of Official Journal (OJ)
citations for decisions associated with the case. Coded \code{NA} if there
are no citations.}

\item{\code{related_cases}}{String. A list of case numbers for related cases,
according to the Commission, separated by a comma. Coded \code{NA} if the
Commission does not identify any related cases.}

\item{\code{count_related_cases}}{Numeric. The number of related cases,
according to the Commission.}

}
}
\usage{
cases
}
\description{
A dataset of all state aid cases opened by the European Commission against EU
member states from 2000 through 2020 under Article 108 of the Treaty on the
Function of the European Union (TFEU) and Council Regulation (EC) No
659/1999, updated by Council Regulation (EU) 2015/1589. These regulations
have a legal basis in Article 109 TFEU.
}
\keyword{datasets}
