% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_csts_dp.R
\docType{data}
\name{cases_csts_dp}
\alias{cases_csts_dp}
\title{CSTS data on cases (by department)}
\format{
A data frame with 6 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the case was opened by the Commission.}
\item{department_id}{Numeric. An ID number that uniquely identifies each
Directorate-General (DG) of the Commission. There are three DGs that can
open state aid cases. The DG for Competition (COMP) is coded \code{1}, the
DG for Agriculture and Rural Development (AGRI) is coded \code{2}, and the
DG for Maritime Affairs and Fisheries (MARE) is coded \code{3}.}
\item{department}{String. The name of the Directorate-General (DG) of the
Commission that opened the state aid case.}
\item{department_code}{String. A multi-letter code assigned by the
Commission that uniquely identifies each Directorate-General (DG) of the
Commission. }
\item{count_cases}{Numeric. A count of the number of cases opened by the
Commission at this level of aggregation.}
}
}
\usage{
cases_csts_dp
}
\description{
This dataset includes aggregated data on the number of state aid cases per
department per year (cross-sectional time-series data). There is one
observation per department per year (1988-2020). The dataset uses current
department names.
}
\keyword{datasets}
