% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_csts_ms_ct.R
\docType{data}
\name{cases_csts_ms_ct}
\alias{cases_csts_ms_ct}
\title{CSTS data on cases (by member state and by case type)}
\format{
A data frame with 8 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the case was opened by the Commission.}
\item{member_state_id}{Numeric. An ID number that uniquely identifies each
member state. This ID number is assigned when member states are sorted by
accession date and then alphabetically. }
\item{member_state}{String. The name of the member state that the
Commission opened the case against. }
\item{member_state_code}{String. A two letter code assigned by the
Commission that uniquely identifies each member state. }
\item{case_type_id}{Numeric. An ID number that uniquely identifies each
type of state aid cases. Coded \code{1} for cases that involve state aid
measures granted through Commission-approved schemes that is not
notifiable, coded \code{2} for cases that involve state aid measures
granted through Commission-approved schemes where the Commission has
required that the member state notify any aid granted through the scheme,
and coded \code{3} for cases that involve ad hoc state aid measures that
are not granted through a scheme.}
\item{case_type}{String. The type of the state aid case. There are three
types of cases. Coded \code{Scheme} for cases that involve state aid
measures granted through Commission-approved schemes that is not
notifiable, coded \code{Individual application} for cases that involve
state aid measures granted through Commission-approved schemes where the
Commission has required that the member state notify any aid granted
through the scheme, and coded \code{Ad hoc} for cases that involve ad hoc
state aid measures that are not granted through a scheme.}
\item{count_cases}{Numeric. A count of the number of cases opened by the
Commission at this level of aggregation.}
}
}
\usage{
cases_csts_ms_ct
}
\description{
This dataset includes aggregated data on the number of state aid cases per
member state per year (cross-sectional time-series data) broken down by
case type. There is one observation per member state per year per case type
(1988-2020), excluding state-years where the state was not a member of the
EU.
}
\keyword{datasets}
