% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_ddy.R
\docType{data}
\name{cases_ddy}
\alias{cases_ddy}
\title{DDY data on cases}
\format{
A data frame with 9 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the case was opened by the Commission.}
\item{department_id}{Numeric. An ID number that uniquely identifies each
Directorate-General (DG) of the Commission. There are three DGs that can
open state aid cases. The DG for Competition (COMP) is coded \code{1}, the
DG for Agriculture and Rural Development (AGRI) is coded \code{2}, and the
DG for Maritime Affairs and Fisheries (MARE) is coded \code{3}.}
\item{department}{String. The name of the Directorate-General (DG) of the
Commission that opened the state aid case.}
\item{department_code}{String. A multi-letter code assigned by the
Commission that uniquely identifies each Directorate-General (DG) of the
Commission. }
\item{member_state_id}{Numeric. An ID number that uniquely identifies each
member state. This ID number is assigned when member states are sorted by
accession date and then alphabetically. }
\item{member_state}{String. The name of the member state that the
Commission opened the case against. }
\item{member_state_code}{String. A two letter code assigned by the
Commission that uniquely identifies each member state. }
\item{count_cases}{Numeric. A count of the number of cases opened by the
Commission at this level of aggregation.}
}
}
\usage{
cases_ddy
}
\description{
This dataset includes aggregated data on the number of state aid cases per
department per member state per year (directed dyad-year data). There is
one observation per department per member state per year (1988-2020),
excluding directed dyad-years where the state was not a member of the EU.
The dataset uses current department names.
}
\keyword{datasets}
