% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_net.R
\docType{data}
\name{cases_net}
\alias{cases_net}
\title{Case-level network data}
\format{
A data frame with 7 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{\code{numeric}. The year the case was opened by the
Commission.}
\item{from_node_id}{\code{numeric}. An ID number that uniquely identifies
each node in the network that creates a link, which is always a Commission
department.}
\item{from_node}{\code{string}. The name of the Commission department that
opened the case.}
\item{to_node_id}{\code{numeric}. An ID number that uniquely identifies
each node in the network that receives a link, which is always a member
state.}
\item{to_node}{\code{string}. The name of the member state that the
Commission opened the case against. }
\item{edge_weight}{\code{numeric}. The weight of the edge, which is the
number of cases opened by the Commission.}
}
}
\usage{
cases_net
}
\description{
This dataset includes network data for state aid cases. Network data is
similar to directed dyad-year data except that it only includes directed
dyad-years with at least one infringement case. For every year, there is
one node per department and one node per member state. Edges can only exist
between a department and a member state. There is an edge between a
department and a member state if and only if the department opened at least
one case against the member state during that year. The weight of the edge
is the number of cases that the department opened against the member state.
There is one observation per department per member state per year
(1988-2020), excluding directed dyad-years where the state was not a member
of the EU, but only if count of cases is positive.
}
\keyword{datasets}
