% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_net_ct.R
\docType{data}
\name{cases_net_ct}
\alias{cases_net_ct}
\title{Network data on cases (by case type)}
\format{
A data frame with 11 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the case was opened by the Commission.}
\item{department_id}{Numeric. An ID number that uniquely identifies each
Directorate-General (DG) of the Commission. There are three DGs that can
open state aid cases. The DG for Competition (COMP) is coded \code{1}, the
DG for Agriculture and Rural Development (AGRI) is coded \code{2}, and the
DG for Maritime Affairs and Fisheries (MARE) is coded \code{3}.}
\item{department}{String. The name of the Directorate-General (DG) of the
Commission that opened the state aid case.}
\item{department_code}{String. A multi-letter code assigned by the
Commission that uniquely identifies each Directorate-General (DG) of the
Commission. }
\item{member_state_id}{Numeric. An ID number that uniquely identifies each
member state. This ID number is assigned when member states are sorted by
accession date and then alphabetically. }
\item{member_state}{String. The name of the member state that the
Commission opened the case against. }
\item{member_state_code}{String. A two letter code assigned by the
Commission that uniquely identifies each member state. }
\item{case_type_id}{Numeric. An ID number that uniquely identifies each
type of state aid cases. Coded \code{1} for cases that involve state aid
measures granted through Commission-approved schemes that is not
notifiable, coded \code{2} for cases that involve state aid measures
granted through Commission-approved schemes where the Commission has
required that the member state notify any aid granted through the scheme,
and coded \code{3} for cases that involve ad hoc state aid measures that
are not granted through a scheme.}
\item{case_type}{String. The type of the state aid case. There are three
types of cases. Coded \code{Scheme} for cases that involve state aid
measures granted through Commission-approved schemes that is not
notifiable, coded \code{Individual application} for cases that involve
state aid measures granted through Commission-approved schemes where the
Commission has required that the member state notify any aid granted
through the scheme, and coded \code{Ad hoc} for cases that involve ad hoc
state aid measures that are not granted through a scheme.}
\item{count_cases}{Numeric. A count of the number of cases opened by the
Commission at this level of aggregation.}
}
}
\usage{
cases_net_ct
}
\description{
This dataset includes multi-dimensional network data for state aid cases.
There is one dimension per case type. Network data is similar to directed
dyad-year data except that it only includes directed dyad-years with at
least one infringement case with respect to each case type. For every year,
there is one node per department and one node per member state. Edges can
only exist between a department and a member state. There is an edge
between a department and a member state with respect to each case type if
and only if the department opened at least one case against the member
state during that year. The weight of the edge is the number of cases that
the department opened against the member state. There is one observation
per department per member state per year per case type (1988-2020),
excluding directed dyad-years where the state was not a member of the EU,
but only if the count of cases is positive.
}
\keyword{datasets}
