% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_ts.R
\docType{data}
\name{cases_ts}
\alias{cases_ts}
\title{Case-level time-series data}
\format{
A data frame with 3 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{\code{numeric}. The year the case was opened by the
Commission.}
\item{count_cases}{\code{numeric}. A count of the number of cases opened by
the Commission at this level of aggregation.}
}
}
\usage{
cases_ts
}
\description{
This dataset includes aggregated data on the number of state aid cases per
year (time-series data). There is one observation per year (1988-2020).
}
\keyword{datasets}
