% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{decisions}
\alias{decisions}
\title{EU State Aid Decisions}
\format{
A tibble with X variables and X observations:

\describe{

\item{\code{key_ID}}{Numeric. An ID number that uniquely identifies all
observations.}

}
}
\usage{
cases
}
\description{
A dataset of all decisions in state aid cases opened by the European Commission
against EU member states from 2000 through 2020.
}
\keyword{datasets}
