% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions_csts_ms.R
\docType{data}
\name{decisions_csts_ms}
\alias{decisions_csts_ms}
\title{Decision-level cross-sectional time-series data by member state}
\format{
A data frame with 8 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{\code{numeric}. The year the case was opened by the
Commission.}
\item{member_state_id}{\code{numeric}. An ID number that uniquely
identifies each member state. This ID number is assigned when member states
are sorted by accession date and then alphabetically. }
\item{member_state}{\code{string}. The name of the member state that the
Commission opened the case against. }
\item{member_state_code}{\code{string}. A two letter code assigned by the
Commission that uniquely identifies each member state. }
\item{decision_type_id}{\code{numeric}. An ID number that uniquely
identifies each type of decision the Commission can make in a state aid
case.}
\item{decision_type}{\code{string}. The type of the decision.}
\item{count_decisions}{\code{numeric}. A count of the number of decisions
issued by the Commission at this level of aggregation.}
}
}
\usage{
decisions_csts_ms
}
\description{
This dataset includes aggregated data on the number of each type of
decision per member state per year (cross-sectional time-series data).
There is one observation per member state per year per decision type
(1988-2020), excluding state-years where the state was not a member of the
EU.
}
\keyword{datasets}
