% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions_ts.R
\docType{data}
\name{decisions_ts}
\alias{decisions_ts}
\title{TS data on decisions}
\format{
A data frame with 5 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the case was opened by the Commission.}
\item{decision_type_id}{Numeric. An ID number that uniquely identifies each
type of decision the Commission can make in a state aid case.}
\item{decision_type}{String. The type of the decision.}
\item{count_decisions}{Numeric. A count of the number of decisions issued
by the Commission at this level of aggregation.}
}
}
\usage{
decisions_ts
}
\description{
This dataset includes aggregated data on the number of each type of
decision per year (time-series data). There is one observation per year per
decision stage (1988-2020).
}
\keyword{datasets}
