% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{EUSA_cases_DDY_API}
\alias{EUSA_cases_DDY_API}
\title{Query Data from the EUSA Database Using the Cases DDY API}
\usage{
EUSA_cases_DDY_API(parameters)
}
\arguments{
\item{parameters}{A named list. The names of the elements of the list should
correspond exactly to parameters of the EUSA Database API. With the
exception of the \code{limit} and \code{offset} parameters, each parameter
corresponds to a variable from the dataset. See the EUSA Database API
Codebook for a list of valid parameters and the corresponding variable in
the dataset, if applicable. See the EUSA Database Codebook for a
description of the variables. All API parameters are in \code{lowerCamelCase}
with no exceptions. With the exception of the \code{limit} and \code{offset}
parameters, value of the elements of the list are the values that will be
used to filter the data. If specified, the \code{limit} parameter indicates the
maximum number of observations to return, and the \code{offset} parameter
indicates which observation of the filtered data to start at. If you
include invalid parameters, the function will still run, but you will get a
warning indicating which parameters are invalid.}
}
\description{
A function to query the \code{cases_DDY}, \code{cases_DDY_D}, \code{cases_network}, and
\code{cases_network_D} datasets of the EUSA Database using the EUSA Database API.
}
