% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_parameters}
\alias{filter_parameters}
\title{A Function to Remove Invalid API Parameters}
\usage{
filter_parameters(provided, valid)
}
\arguments{
\item{provided}{A string vector. The names of the API parameters provided by the
user. These are the names of the elements of the list proved via the
\code{parameters} argument of the parent function.}

\item{valid}{A string vector. The names of all of the valid paramters for a model
of the EUSA Database API.}
}
\description{
A function to remove invalid API paramters from the named list of parameter
values provided by the user.
}
