% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{nace}
\alias{nace}
\title{NACE codes}
\format{
A tibble with 4 variables and 839 observations:
\describe{

\item{\code{key_ID}}{Numeric. A unique identifier for each observation.}

\item{\code{NACE_sector}}{String. The NACE sector for the observation
(letters A to U).}

\item{\code{NACE_code}}{String. The NACE sector, divison, group, or class
for the observation.}

\item{\code{NACE_description}}{String. The NACE description for the sector,
division, group, or class.}

}
}
\usage{
nace
}
\description{
A dataset of NACE codes and descriptions.
}
\details{
The Statistical Classification of Economic Activities in the European
Community (NACE, after the French "nomenclature statistique des activités
économiques dans la Communauté européenne") is the EU's industry standard
classification system (see Regulation (EC) No 1893/2006). NACE codes have
four levels: \itemize{\item Level 1: 21 sectors with letters (A to U) \item
Level 2: 88 divisions with two-digit codes (01 to 99) \item Level 3: 72
groups with three-digit codes (01.1 to 99.0) \item Level 4: 615 classes with
four-digit codes (01.11 to 99.00)}
}
\keyword{datasets}
