% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{query_API}
\alias{query_API}
\title{Query the EUSA Database API}
\usage{
query_API(parameters, route)
}
\arguments{
\item{parameters}{A named list. The names of the elements of the list should
correspond exactly to parameters of the EUSA Database API. With the
exception of the \code{limit} and \code{offset} parameters, each parameter
corresponds to a variable from the dataset. See the EUSA Database API
Codebook for a list of valid parameters and the corresponding variable in
the dataset, if applicable. See the EUSA Database Codebook for a
description of the variables. All API parameters are in \code{lowerCamelCase}
with no exceptions. With the exception of the \code{limit} and \code{offset}
parameters, value of the elements of the list are the values that will be
used to filter the data. If specified, the \code{limit} parameter indicates the
maximum number of observations to return, and the \code{offset} parameter
indicates which observation of the filtered data to start at. If you
include invalid parameters, the function will still run, but you will get a
warning indicating which parameters are invalid.}

\item{route}{A string. The name of the API route to query.}
}
\description{
A function to query the EUSA Database API.
}
