% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{awards}
\alias{awards}
\title{EU State Aid Awards}
\format{
A tibble with 22 variables and 92,968 observations:

\describe{

\item{\code{key_ID}}{Numeric. An ID number that uniquely identifies all
observations.}

\item{\code{case_number}}{String. An ID number that uniquely identifies the
Commission state aid case assocaited with the award.}

\item{\code{reference_number}}{String. An ID number that uniquely identifies
each state aid award.}

\item{\code{notification_date}}{String. The date the member state notified
the award to the Comission under the transparency requirements in the format
\code{YYYY-MM-DD}.}

\item{\code{publication_date}}{String. The date the Commission published the
award in the transparency database.}

\item{\code{member_state}}{String. The name of the member state that notified
the award.}

\item{\code{authority}}{String. The name of the national, regional, or local
government authority that granted the state aid award. Note that this
variable is not cleaned. The text of the names appears exactly how it is
entered in the transparency database.}

\item{\code{region}}{String. The name of the region where the state aid award
was granted. This variable uses the Nomenclature des Unités Territoriales
Statistiques (NUTS) classification system for subnational territories. Note
that this variable is not cleaned. The text of the names appears exactly how
it is entered in the transparency database.}

\item{\code{beneficiary_name}}{String. The name of the firm that recieved the
state aid award. Note that this variable is not cleaned. The text of the
names appears exactly how it is entered in the transparency database.}

\item{\code{beneficiary_type}}{String. The type of the firm that received the
state aid award. Possible values include \code{small and medium-sized
enterprise (SME)} and \code{large enterprise}}.

\item{\code{NACE_sector}}{String. The NACE sector of the firm that received the state aid award.
NACE sectors are identified by the letters A to U.}

\item{\code{NACE_code}}{String. The NACE sector, division, group, or class of the firm that received
the state aid award. Note that some firms are coded at a higher level of specificity than others in the transparency database.}

\item{\code{NACE_description}}{String. The NACE description of the sector, division, group, or class of the firm
that received the state aid award.}

\item{\code{aid_instrument}}{The policy instrument used in granting the
award. Possible values include: \itemize{
\item{\code{direct grant} — Definition.}
\item{\code{direct grant/interest rate subsidy} — Definition.}
\item{\code{guarantee} — Definition.}
\item{\code{interest subsidy} — Definition.}
\item{\code{loan/repayable advances} — Definition.}
\item{\code{other} — Definition.}
\item{\code{provision of risk capital} — Definition.}
\item{\code{provision of risk finance} — Definition.}
\item{\code{reduction of social security contributions} — Definition.}
\item{\code{reimbursable grant} — Definition.}
\item{\code{repayable advances} — Definition.}
\item{\code{soft loan} — Definition.}
\item{\code{subsidised services} — Definition.}
\item{\code{tax advantage or tax exemption} — Definition.}
\item{\code{tax allowance} — Definition.}
\item{\code{tax base reduction} — Definition.}
\item{\code{tax rate reduction} — Definition.}
}}

\item{\code{raw_amount}}{Numeric. The amount of the award denominated in the
currency of the member state that provided the award. Coded \code{NA} if the
amount of the award is not recorded in the transparency database. This is the
actual or estimated economic value of the award as reported by the member
state, not only the nominal amount. When reporting the value of the award,
member states can report a range if the exact value would reveal valuable
information to the competitors of the beneficiary. When this occurs, the
value of this variable is the midpoint of the reported range. Note that
values are not comparable across observations because different observations
are denominated in different currencies.}

\item{\code{currency}}{String. The currency that the raw amount of the award
is denominated in. Coded \code{NA} if the amount of the award is not reported
in the transparency database. Possible values include: \itemize{
\item{\code{euro} (EUR, European Union)}
\item{\code{forint} (HUF, Hungary)}
\item{\code{koruna} (CZK, Czech Republic)}
\item{\code{krona} (SEK, Sweden)}
\item{\code{krone} (DKK, Denmark)}
\item{\code{kuna} (HRK, Croatia)}
\item{\code{lev} (BGN, Bulgaria)}
\item{\code{pound} (GBP, United Kingdom)}
}}

\item{\code{range}}{Numeric. A dummy variable indicating whether the member
state reported the amount of the award as a range. Coded \code{NA} if the
amount of the award is not reported in the transparency database.}

\item{\code{range_min}}{Numeric. If the member state reported the amount of the award
as a range, the minimum value of the award. Coded \code{NA} if the member
state did not report the amount as the range or if the amount of the award is
not reported in the transparency database.}

\item{\code{range_max}}{Numeric. If the member state reported the amount of the award
as a range, the maximum value of the award. Coded \code{NA} if the member
state did not report the amount as the range or if the amount of the award is
not reported in the transparency database.}

\item{\code{exchange_rate}}{Numeric. The value of one unit of the currency of the
member state that granted the award in euros on the weekday to the
notificate date that a exchange rate quotation is available. Used to convert
the raw amount of the state aid award to euros. This is an exchange rate
quotation in which the currency of the member state is the fixed currency and
the euro is the variable currency. For example, if the currency in question
is the pound sterling (GBR), and the GBR/EUR exchange rate is approximately
1.2, then one pound is equivalent to 1.2 euros. If the value of the award was
reported in euros, the value will be 1. The daily exchange rate quotations
are from Yahoo Finance.}

\item{\code{adjusted_amount}}{Numeric. The amount of the award in euros. Coded
\coded{NA}. This is the raw value of the award in the currency of the member
state that granted the award times the exchange rate (the value of one unit
of that currency in euros). Unlike the raw amount, this variable is directly
comparable across observations.}

\item{\code{voluntary}}{Numeric. A dummy variable indicating whether the amount
of the state aid award is under the manditory reporting theshold.}

}
}
\usage{
cases
}
\description{
A dataset of all awards granted by EU member states to firms as part of
approved state aid schemes from 2016 through 2020. The unit of observation is
an individual state aid award granted through a state aid scheme already
approved by the Commission or as an ad hoc state aid measure approved by the
Commission.
}
\details{
The State Aid Modernisation programme (SAM) was adopted in 2014
and came into force on 1 July 2016.
}
\keyword{datasets}
