% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions.R
\docType{data}
\name{decisions}
\alias{decisions}
\title{Decision-level data}
\format{
A data frame with 30 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation in the dataset. }
\item{case_id}{\code{string}. An ID number that uniquely identifies each
state aid case. Assigned by the Commission. The Commission changed the
format of case numbers in 2010. Before this change, the case number
indicated the type(s) of the procedure associated with the case. After this
change, all case numbers have the format \code{SA.#####}.}
\item{procedure_numbers}{\code{string}. Numbers that uniquely identifiy
each procedure associated with a case. If there are multiple procedures,
the procedure numbers are listed and separated by a comma. After the
Commission changed the format of case numbers in 2010, only procedure
numbers indicate the types of procedures associated with a case.}
\item{member_state_id}{\code{numeric}. An ID number that uniquely
identifies each member state. This ID number is assigned when member states
are sorted by accession date and then alphabetically. }
\item{member_state}{\code{string}. The name of the member state that the
Commission opened the case against. }
\item{member_state_code}{\code{string}. A two letter code assigned by the
Commission that uniquely identifies each member state. }
\item{department_id}{\code{numeric}. An ID number that uniquely identifies
each Directorate-General (DG) of the Commission. There are three DGs that
can open state aid cases. The DG for Competition (COMP) is coded \code{1},
the DG for Agriculture and Rural Development (AGRI) is coded \code{2}, and
the DG for Maritime Affairs and Fisheries (MARE) is coded \code{3}.}
\item{department}{\code{string}. The name of the Directorate-General (DG)
of the Commission that opened the state aid case.}
\item{department_code}{\code{string}. A multi-letter code assigned by the
Commission that uniquely identifies each Directorate-General (DG) of the
Commission. }
\item{case_type_id}{\code{numeric}. An ID number that uniquely identifies
each type of state aid cases. Coded \code{1} for cases that involve state
aid measures granted through Commission-approved schemes that is not
notifiable, coded \code{2} for cases that involve state aid measures
granted through Commission-approved schemes where the Commission has
required that the member state notify any aid granted through the scheme,
and coded \code{3} for cases that involve ad hoc state aid measures that
are not granted through a scheme.}
\item{case_type}{\code{string}. The type of the state aid case. There are
three types of cases. Coded \code{Scheme} for cases that involve state aid
measures granted through Commission-approved schemes that is not
notifiable, coded \code{Individual application} for cases that involve
state aid measures granted through Commission-approved schemes where the
Commission has required that the member state notify any aid granted
through the scheme, and coded \code{Ad hoc} for cases that involve ad hoc
state aid measures that are not granted through a scheme.}
\item{procedure_types}{\code{string}. Numbers that uniquely identified each
procedure associated with a case. If there are multiple procedures, the
procedure numbers are listed and separated by a comma. After the Commission
changed the format of case numbers in 2010, only procedure numbers indicate
the types of procedures associated with a case.}
\item{contradictory_aid}{\code{dummy}. A dummy variable indicating whether
the case relates to whether the Commission suspects that the state aid
measure is contradictory to the rules of the single market, as defined by
Article 107 of the Treaty of the Functioning of the European Union (TFEU).
Coded \code{1} if it does and \code{0} if it does not.}
\item{existing_aid}{\code{dummy}. A dummy variable indicating whether the
case relates to existing state aid measures. Coded \code{1} if it does and
\code{0} if it does not.}
\item{notified_aid}{\code{dummy}. A dummy variable indicating whether the
case relates to state aid notified by the member state. Coded \code{1} if
it does and \code{0} if it does not.}
\item{unnotified_aid}{\code{dummy}. A dummy variable indicating whether the
case relates to state aid that is not notified by the member state. Coded
\code{1} if it does and \code{0} if it does not.}
\item{general_block_exemption}{\code{dummy}. A dummy variable indicating
whether the case relates to state that is exempt from notification under
the General Block Exemption Regulation (GBER). Coded \code{1} if it does
and \code{0} if it does not.}
\item{specific_block_exemption}{\code{dummy}. A dummy variable indicating
whether the case relates to state aid that is exempt from notification
under a specific block exemption. Coded \code{1} if it does and \code{0} if
it does not.}
\item{notification_date}{\code{date}. The date the member state notified
the Commission of the state aid measure in the format \code{YYYY-MM-DD}. }
\item{notification_year}{\code{numeric}. The year the member state notified
the Commission of the state aid measure.}
\item{notification_month}{\code{numeric}. The month the member state
notified the Commission of the state aid measure.}
\item{notification_day}{\code{numeric}. The day the member state notified
the Commission of the state aid measure.}
\item{outcome_date}{\code{date}. The date the Commission took its last
decision in the case in the format \code{YYYY-MM-DD}.}
\item{outcome_year}{\code{numeric}. The year the Commission took its last
decision in the case.}
\item{outcome_month}{\code{numeric}. The month the Commission took its last
decision in the case.}
\item{outcome_day}{\code{numeric}. The day the Commission took its last
decision in the case.}
\item{decision_number}{\code{numeric}. The number of the decision within
each case.}
\item{decision_type_id}{\code{numeric}. An ID number that uniquely
identifies each type of decision the Commission can make in a state aid
case.}
\item{decision_type}{\code{string}. The type of the decision.}
\item{phase}{\code{string}. The phase of the procedure. Coded \code{Exempt}
for state aid that is exempt from notification, \code{Phase 1} for
decisions made during a preliminary investigation, \code{Phase 2} for
decisions made during a formal investigation, or \code{Other} for decisions
that the Commission can make during either phase. }
}
}
\usage{
decisions
}
\description{
This dataset include data on decisions in state aid cases. There is one
observation per decision per case (1988-2020). The dataset includes
information on the department responsible for the decision and the member
state that the decision is against.
}
\keyword{datasets}
