% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{decisions}
\alias{decisions}
\title{EU State Aid Decisions}
\format{
A tibble with X variables and X observations:

\describe{

\item{\code{key_ID}}{Numeric. An ID number that uniquely identifies all
observations.}

\item{\code{case_number}}{String. An ID number that uniquely identifies each
case.}

\item{\code{procedure_number}}{String. A list of ID numbers that identify
the proceedures associated with the case, separated by a comma.}

\item{\code{member_state}}{String. The name of the member state that notified
that state aid measure.}

\item{\code{directorate_general}}{String. The name of the Directorate-General
(DG) of the Commission that opened the case against the member state.
Possible values include: \itemize{
\item{\code{DG Agriculture and Rural Development}}
\item{\code{DG Competition}}
\item{\code{DG Maritime Affairs and Fisheries}}
}}

\item{\code{directorate_general_code}}{String. The alphabetical code for the
Directorate-General (DG) of the Commission that opened the case against the
member state. Possible values include: \itemize{
\item{\code{AGRI} (for DG Agriculture and Rural Development)}
\item{\code{COMP} (for DG Competition)}
\item{\code{MARE} (for DG Maritime Affairs and Fisheries)}
}}

\item{\code{case_type}}{String. The type of the case. Coded \code{NA} if no
case type is recorded in the database. Possible values include:
\itemize{\item{\code{ad hoc case} — A case involving a state aid measure that
was not granted as part of a scheme that was already approved by the
Commission. } \item{\code{individual application} — A case involving a state
aid measure that was granted as part of a scheme that was already approved by
the Commission (i.e., an individual application of the scheme). }
\item{\code{scheme} — A cases involving a state aid scheme where a member
state can grant awards without each award being approved individually by the
Commission.}}}

\item{\code{notification_date}}{String. The date that the member state
responsible for the state aid mesasure notified the measure to the Commission
under Article 108(3) TFEU in the format \code{YYYY-MM-DD.}}

\item{\code{decision_ID}}{Numeric. An ID number that uniquely identifies each
decision within a case.}

\item{\code{decisions}}{String. A description of the decision made by the
Commission. Possible values include: \itemize{
\item{\code{Article 108(4) of the Treaty on the Functioning of the European
Union (TFEU): exempt from notification under Article 108(3)} — The state aid
measure notified by the member state is exempt from notification under
Article 108(3) TFEU, either because the measure is exempt under the General
Block Exemption Regulation (GBER) (Commission Regulation (EC) No 800/2008),
or because the aid is de minimis.}
\item{\code{Article 4(2) of Council Regulation (EC) No 659/1999: decision
does not constitute aid} — After a preliminary examination, Commission
decided that the state aid measure notified by the member state does not
constitute state aid under Article 107 TFEU.}
\item{\code{Article 4(3) of Council Regulation (EC) No 659/1999: decision not
to raise objections} — After a preliminary examination, Commission decided
not to raise objections to the state aid measure notified by the member state
because the measure is compatible with the rules of the single market, as
defined by Article 107 TFEU.}
\item{\code{Article 4(4) of Council Regulation (EC) No 659/1999: decision to
initiate the formal investigation procedure} — After a preliminary
examination the Comission decided to initiate a formal investigation to
determine whether the state aid measure notified by the member state is
compatible with the rules of the single market, as defined by Article 107
TFEU.}
\item{\code{Article 4(4) of Council Regulation (EC) No 659/1999: decision to
extend proceedings} — The Commission decided to extend the two-month
preliminary examination phase of the state aid procedure.}
\item{\code{Article 7(2) of Council Regulation (EC) No 659/1999: decision
does not constitute aid (after formal investigation procedure)} — After a
formal investigation, the Commission decided that the state aid measure
notified by the member state does not constitute state aid under Article 107
TFEU.}
\item{\code{Article 7(3) of Council Regulation (EC) No 659/1999: positive
decision} — After a formal investigation, the Commission decided that the
state aid measure notified by the member state is compatible with the rules
of the single market, as defined by Article 107 TFEU, and approved the
measure.}
\item{\code{Article 7(4) of Council Regulation (EC) No 659/1999: conditional
decision} — After a formal investigation, the Commission decided to approve
the state aid mesasure notified by the member state subject to conditions to
make the measure compatible with the rules of the single market, as defined
by Article 107 TFEU.}
\item{\code{Articles 7(5) and 14(1) of Council Regulation (EC) No 659/1999:
negative decision with recovery} — After a formal investigation, the
Commission decided that the state aid measure notified by the member state is
not compatible with the rules of the single market, as defined by Article 107
TFEU, and decided that the member state must take all necessary measures to
recover the aid from the beneficiary.}
\item{\code{Articles 7(5) and 14(1) of Council Regulation (EC) No 659/1999:
negative decision with recovery} — After a formal investigation, the
Commission decided that the state aid measure notified by the member state is
not compatible with the rules of the single market, as defined by Article 107
TFEU, and did not approve the measure.}
\item{\code{Article 7(5) of Council Regulation (EC) No 659/1999: negative
decision on notified aid} — After a formal investigation, the Commission
decided that the state aid measure notified by the member state is not
compatible with the rules of the single market, as defined by Article 107
TFEU.}
\item{\code{Article 8(1) of Council Regulation (EC) No 659/1999: withdrawal
of notification (before formal investigation procedure)} — The member state
that notified the state aid measure decided to withdraw the notification
before the formal investigation procedure.}
\item{\code{Article 8(2) of Council Regulation (EC) No 659/1999: withdrawal
of notification (after formal investigation procedure)} — The member state
that notified the state aid measure decided to withdraw the notification
after the formal investigation procedure.}
\item{\code{Article 9 of Council Regulation (EC) No 659/1999: revocation of a
decision} — The Commission decided to revoke a previous decision because
information provided by the member state that the Commisison used in making
the decision was incorrect.}
\item{\code{Article 10(3) of Council Regulation (EC) No 659/1999: information
injunction} — The Commission decided to require the member state to provide
additional information after the member state did not provide information
that the Commission requested or provided incomplete information.}
\item{\code{Article 11(1) of Council Regulation (EC) No 659/1999: suspension
injunction} — The Commission decided to require the member state to suspend
the state aid measure until the Commission has decided whether the measure is
compatible with the rules of the single market, as defined by Article 107
TFEU.}
\item{\code{Article 18 of Council Regulation (EC) No 659/1999: proposal for
appropriate measures} — The Commission decided to issue a recommendation
proposing appropriate measures to be taken by the member state because the
state aid scheme is not compatible with the rules of the single market, as
defined by Article 107 TFEU.}
\item{\code{Article 260(2) (ex Article 228(2)) of the Treaty on the
Functioning of the European Union (TFEU): referral to the Court of Justice
(non-compliance with a Court judgment)} — The Commission decided to refer the
case to the Court of Justice for non-compliance with a Court judgment.}
\item{\code{Article 108(2) of the Treaty on the Functioning of the European
Union (TFEU): referral to the Court of Justice (non-compliance of a
Commission decision)} — The Commission decided to refer the case to the Court
of Justice for non-implementation of a Commission decision issued under
Article 108(2) TFEU.}
\item{\code{Article 12 of Council Regulation (EC) No 659/1999: referral to
the Court of Justice (non-compliance with an injunction)} — The Commission
decided to refer the case to the Court of Justice for non-compliance with an
injunction issued under Article 10(3) or Article 11(1) of Council Regulation
(EC) No 659/1999.}
}}

}
}
\usage{
decisions
}
\description{
This dataset includes all decisions made in all state aid cases under Article
108(2) TFEU and Council Regulation (EU) 2015/1589 over the same period. The
unit of observation is a decision. The source of the raw data is the
Commission's state aid cases database.
}
\keyword{datasets}
